flask/requirements.txt
Flask
flask-socketio
kafka-python>=2.0.2
pymongo
pytz